pyproject.toml
[project]
name = "iocshuttle"
version = "0.1.0"
description = "A small python tool for continually importing IoCs from a TAXII feed into various locations (EDL lists, Crowdstrike, etc.)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "stix-shifter>=7.0.11",
    "taxii2-client>=2.3.0",
    "whenever>=0.6.7",
    "stix-shifter-utils>=7.0.11",
    "crowdstrike-falconpy>=1.4.5",
    "pip-system-certs>=4.0",
    "netaddr>=1.3.0",
    "schedule>=1.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
"ioc" = "ioc:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ioc"]

[tool.uv]
dev-dependencies = [
    "pyinstaller>=6.10.0",
    "setuptools>=78.1.0",
]
override-dependencies = [
    "numpy>=1.26"
]